feat: introduce new  prop for customizing the feed header
